requirements.txt
fastapi
python-multipart
opencv-python-headless
pytesseract
uvicorn
pillow